int to double conversion

+++ Assignment1/src/Q10.java
@@ -6,7 +6,7 @@ public class Q10 {
         double [][] arr = new double[3][4];
         for (int i = 0; i < 3; i++) {
             for (int j = 0; j < 4; j++)
-                arr[i][j] = hv.nextInt();
+                arr[i][j] = hv.nextDouble();
         }
 
         System.out.println("The following 2D Matrix");
